ChannelFactory implementation to use SharedChannelRegistry
Create ChannelFactory interface
