+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/binding/ChannelFactory.java
@@ -15,252 +15,23 @@
  */
 package org.springframework.cloud.stream.binding;
 
-import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.BeanFactory;
-import org.springframework.beans.factory.BeanFactoryAware;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
-import org.springframework.cloud.stream.aggregate.SharedChannelRegistry;
-import org.springframework.cloud.stream.annotation.Input;
-import org.springframework.cloud.stream.annotation.Output;
-import org.springframework.cloud.stream.binder.MessageChannelBinderSupport;
-import org.springframework.cloud.stream.config.BindingProperties;
-import org.springframework.cloud.stream.config.ChannelBindingServiceProperties;
-import org.springframework.cloud.stream.converter.AbstractFromMessageConverter;
-import org.springframework.cloud.stream.converter.ByteArrayToStringMessageConverter;
-import org.springframework.cloud.stream.converter.CompositeMessageConverterFactory;
-import org.springframework.cloud.stream.converter.JavaToSerializedMessageConverter;
-import org.springframework.cloud.stream.converter.JsonToPojoMessageConverter;
-import org.springframework.cloud.stream.converter.JsonToTupleMessageConverter;
-import org.springframework.cloud.stream.converter.MessageConverterUtils;
-import org.springframework.cloud.stream.converter.PojoToJsonMessageConverter;
-import org.springframework.cloud.stream.converter.PojoToStringMessageConverter;
-import org.springframework.cloud.stream.converter.SerializedToJavaMessageConverter;
-import org.springframework.cloud.stream.converter.StringToByteArrayMessageConverter;
-import org.springframework.cloud.stream.converter.TupleToJsonMessageConverter;
-import org.springframework.core.annotation.AnnotationUtils;
-import org.springframework.integration.channel.AbstractMessageChannel;
-import org.springframework.integration.channel.DirectChannel;
-import org.springframework.integration.channel.QueueChannel;
-import org.springframework.integration.config.ConsumerEndpointFactoryBean;
-import org.springframework.integration.scheduling.PollerMetadata;
 import org.springframework.messaging.MessageChannel;
-import org.springframework.messaging.PollableChannel;
-import org.springframework.messaging.SubscribableChannel;
-import org.springframework.messaging.converter.MessageConverter;
-import org.springframework.scheduling.support.PeriodicTrigger;
-import org.springframework.util.Assert;
-import org.springframework.util.MimeType;
-import org.springframework.util.ReflectionUtils;
-import org.springframework.util.StringUtils;
 
 /**
- * Class that {@link BindableProxyFactory} uses to create message channels.
+ * Defines methods to create/configure the {@link org.springframework.messaging.MessageChannel} defined
+ * in {@link org.springframework.cloud.stream.annotation.EnableBinding}.
  *
- * @author Marius Bogoevici
- * @author David Syer
  * @author Ilayaperumal Gopinathan
  */
-public class ChannelFactory implements BeanFactoryAware, InitializingBean {
+public interface ChannelFactory {
 
-	public static final String SPRING_CLOUD_STREAM_INTERNAL_PREFIX = "spring.cloud.stream.internal";
+	String SPRING_CLOUD_STREAM_INTERNAL_PREFIX = "spring.cloud.stream.internal";
 
-	public static final String CHANNEL_NAMESPACE_PROPERTY_NAME = SPRING_CLOUD_STREAM_INTERNAL_PREFIX + ".channelNamespace";
+	String POLLABLE_BRIDGE_INTERVAL_PROPERTY_NAME = SPRING_CLOUD_STREAM_INTERNAL_PREFIX + ".pollableBridge.interval";
 
-	public static final String POLLABLE_BRIDGE_INTERVAL_PROPERTY_NAME = SPRING_CLOUD_STREAM_INTERNAL_PREFIX + ".pollableBridge.interval";
+	String CHANNEL_NAMESPACE_PROPERTY_NAME = SPRING_CLOUD_STREAM_INTERNAL_PREFIX + ".channelNamespace";
 
-	@Value("${" + CHANNEL_NAMESPACE_PROPERTY_NAME + ":}")
-	private String channelNamespace;
+	MessageChannel createAndConfigureMessageChannel(String name, Class<?> channelType) throws Exception;
 
-	@Value("${" + POLLABLE_BRIDGE_INTERVAL_PROPERTY_NAME + ":1000}")
-	private int pollableBridgeDefaultFrequency;
-
-	private ConfigurableListableBeanFactory beanFactory;
-
-	@Autowired(required = false)
-	private SharedChannelRegistry sharedChannelRegistry;
-
-	private CompositeMessageConverterFactory messageConverterFactory;
-
-	private final ChannelBindingServiceProperties channelBindingServiceProperties;
-
-	public ChannelFactory(ChannelBindingServiceProperties channelBindingServiceProperties) {
-		this.channelBindingServiceProperties = channelBindingServiceProperties;
-	}
-
-	@Override
-	public void setBeanFactory(BeanFactory beanFactory) throws BeansException {
-		this.beanFactory = (ConfigurableListableBeanFactory) beanFactory;
-	}
-
-	@Override
-	public void afterPropertiesSet() throws Exception {
-		Assert.notNull(this.beanFactory, "Bean factory cannot be empty");
-		Set<AbstractFromMessageConverter> messageConverters = new HashSet<>();
-		messageConverters.add(new JsonToTupleMessageConverter());
-		messageConverters.add(new TupleToJsonMessageConverter());
-		messageConverters.add(new JsonToPojoMessageConverter());
-		messageConverters.add(new PojoToJsonMessageConverter());
-		messageConverters.add(new ByteArrayToStringMessageConverter());
-		messageConverters.add(new StringToByteArrayMessageConverter());
-		messageConverters.add(new PojoToStringMessageConverter());
-		messageConverters.add(new JavaToSerializedMessageConverter());
-		messageConverters.add(new SerializedToJavaMessageConverter());
-		this.messageConverterFactory = new CompositeMessageConverterFactory(messageConverters);
-	}
-
-	ChannelHolder createInputMessageChannel(Class<?> type) throws Exception {
-		final List<ChannelHolder> channelHolderList = new ArrayList<>();
-		ReflectionUtils.doWithMethods(type, new ReflectionUtils.MethodCallback() {
-			@Override
-			public void doWith(Method method) throws IllegalArgumentException {
-
-				Input input = AnnotationUtils.findAnnotation(method, Input.class);
-				if (input != null) {
-					String name = BindingBeanDefinitionRegistryUtils.getChannelName(input, method);
-					Class<?> inputChannelType = method.getReturnType();
-					MessageChannel sharedChannel = locateSharedChannel(name);
-					if (sharedChannel == null) {
-						MessageChannel inputChannel = createMessageChannel(inputChannelType);
-						configureMessageConverters(inputChannel, name);
-						channelHolderList.add(new ChannelHolder(name, inputChannel, true));
-					}
-					else {
-						if (inputChannelType.isAssignableFrom(sharedChannel.getClass())) {
-							configureMessageConverters(sharedChannel, name);
-							channelHolderList.add(new ChannelHolder(name, sharedChannel, false));
-						}
-						else {
-							// handle the special case where the shared channel is of a different nature
-							// (i.e. pollable vs subscribable) than the target channel
-							final MessageChannel inputChannel = createMessageChannel(inputChannelType);
-							if (isPollable(sharedChannel.getClass())) {
-								bridgePollableToSubscribableChannel(sharedChannel, inputChannel);
-							}
-							else {
-								bridgeSubscribableToPollableChannel((SubscribableChannel) sharedChannel, inputChannel);
-							}
-							configureMessageConverters(inputChannel, name);
-							channelHolderList.add(new ChannelHolder(name, inputChannel, false));
-						}
-					}
-				}
-			}
-		});
-		return (channelHolderList.isEmpty()) ? null : channelHolderList.get(0);
-	}
-
-	ChannelHolder createOutputMessageChannel(Class<?> type) throws Exception {
-		final List<ChannelHolder> channelHolderList = new ArrayList<>();
-		ReflectionUtils.doWithMethods(type, new ReflectionUtils.MethodCallback() {
-			@Override
-			public void doWith(Method method) throws IllegalArgumentException {
-
-				Output output = AnnotationUtils.findAnnotation(method, Output.class);
-				if (output != null) {
-					String name = BindingBeanDefinitionRegistryUtils.getChannelName(output, method);
-					Class<?> messageChannelType = method.getReturnType();
-					MessageChannel sharedChannel = locateSharedChannel(name);
-					if (sharedChannel == null) {
-						MessageChannel outputChannel = createMessageChannel(messageChannelType);
-						configureMessageConverters(outputChannel, name);
-						channelHolderList.add(new ChannelHolder(name, outputChannel, true));
-					}
-					else {
-						if (messageChannelType.isAssignableFrom(sharedChannel.getClass())) {
-							configureMessageConverters(sharedChannel, name);
-							channelHolderList.add(new ChannelHolder(name, sharedChannel, false));
-						}
-						else {
-							// handle the special case where the shared channel is of a different nature
-							// (i.e. pollable vs subscribable) than the target channel
-							final MessageChannel outputChannel = createMessageChannel(messageChannelType);
-							if (isPollable(messageChannelType)) {
-								bridgePollableToSubscribableChannel(outputChannel, sharedChannel);
-							}
-							else {
-								bridgeSubscribableToPollableChannel((SubscribableChannel) outputChannel, sharedChannel);
-							}
-							configureMessageConverters(outputChannel, name);
-							channelHolderList.add(new ChannelHolder(name, outputChannel, false));
-						}
-					}
-				}
-			}
-		});
-		return (channelHolderList.isEmpty()) ? null : channelHolderList.get(0);
-	}
-
-	private MessageChannel locateSharedChannel(String name) {
-		return this.sharedChannelRegistry != null ? this.sharedChannelRegistry.get(getNamespacePrefixedChannelName(name)) : null;
-	}
-
-	private String getNamespacePrefixedChannelName(String name) {
-		return this.channelNamespace + "." + name;
-	}
-
-	private void bridgeSubscribableToPollableChannel(SubscribableChannel sharedChannel,
-			MessageChannel inputChannel) {
-		sharedChannel.subscribe(new MessageChannelBinderSupport.DirectHandler(
-				inputChannel));
-	}
-
-	private void bridgePollableToSubscribableChannel(MessageChannel pollableChannel,
-			MessageChannel subscribableChannel) {
-		ConsumerEndpointFactoryBean consumerEndpointFactoryBean = new ConsumerEndpointFactoryBean();
-		consumerEndpointFactoryBean.setInputChannel(pollableChannel);
-		PollerMetadata pollerMetadata = new PollerMetadata();
-		pollerMetadata.setTrigger(new PeriodicTrigger(this.pollableBridgeDefaultFrequency));
-		consumerEndpointFactoryBean.setPollerMetadata(pollerMetadata);
-		consumerEndpointFactoryBean
-				.setHandler(new MessageChannelBinderSupport.DirectHandler(
-						subscribableChannel));
-		consumerEndpointFactoryBean.setBeanFactory(this.beanFactory);
-		try {
-			consumerEndpointFactoryBean.afterPropertiesSet();
-		}
-		catch (Exception e) {
-			throw new IllegalStateException(e);
-		}
-		consumerEndpointFactoryBean.start();
-	}
-
-	private MessageChannel createMessageChannel(Class<?> messageChannelType) {
-		return isPollable(messageChannelType) ? new QueueChannel() : new DirectChannel();
-	}
-
-	/**
-	 * Setup data-type and message converters for the given message channel.
-	 *
-	 * @param channel message channel to set the data-type and message converters
-	 * @param channelName the channel name
-	 */
-	private void configureMessageConverters(MessageChannel channel, String channelName) {
-		Assert.isAssignable(AbstractMessageChannel.class, channel.getClass());
-		AbstractMessageChannel messageChannel = (AbstractMessageChannel) channel;
-		BindingProperties bindingProperties = this.channelBindingServiceProperties.getBindings().get(channelName);
-		if (bindingProperties != null) {
-			String contentType = bindingProperties.getContentType();
-			if (StringUtils.hasText(contentType)) {
-				MimeType mimeType = MessageConverterUtils.getMimeType(contentType);
-				MessageConverter messageConverter = this.messageConverterFactory.newInstance(mimeType);
-				Class<?> dataType = MessageConverterUtils.getJavaTypeForContentType(mimeType,
-						Thread.currentThread().getContextClassLoader());
-				messageChannel.setDatatypes(dataType);
-				messageChannel.setMessageConverter(messageConverter);
-			}
-		}
-	}
-
-	private boolean isPollable(Class<?> channelType) {
-		return PollableChannel.class.equals(channelType);
-	}
+	void configureSharedMessageChannel(String name, Class<?> channelType, MessageChannel sharedChannel);
 }
