+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/config/ChannelBindingServiceConfiguration.java
@@ -32,6 +32,7 @@ import org.springframework.cloud.stream.binding.BinderAwareChannelResolver;
 import org.springframework.cloud.stream.binding.BinderAwareRouterBeanPostProcessor;
 import org.springframework.cloud.stream.binding.ChannelBindingService;
 import org.springframework.cloud.stream.binding.ChannelFactory;
+import org.springframework.cloud.stream.binding.ChannelFactoryImpl;
 import org.springframework.cloud.stream.binding.ContextStartAfterRefreshListener;
 import org.springframework.cloud.stream.binding.InputBindingLifecycle;
 import org.springframework.cloud.stream.binding.OutputBindingLifecycle;
@@ -73,7 +74,7 @@ public class ChannelBindingServiceConfiguration {
 
 	@Bean
 	public ChannelFactory channelFactory(ChannelBindingServiceProperties channelBindingServiceProperties) {
-		return new ChannelFactory(channelBindingServiceProperties);
+		return new ChannelFactoryImpl(channelBindingServiceProperties);
 	}
 
 	@Bean
